refactor(enterprise): fold relay dialer into entchatd, delete chats.go
Remove the RemotePartsProvider indirection entirely. The WebSocket
relay dial logic (newRemotePartsProvider, relayHeaderTokenProvider,
relayHeaders, RelaySourceHeader) moves into the enterprise/coderd/chatd
package as the unexported dialRelay function.

NewMultiReplicaSubscribeFn now takes a Config struct with:
- ResolveReplicaAddress: replica URL resolver
- ReplicaHTTPClient: mesh-TLS HTTP client
- ReplicaIDFn: lazy replica ID (not known at init)
- DialerFn: optional override for tests

coderd.go wiring simplifies from a 25-line anonymous closure to a
3-field Config literal. enterprise/coderd/chats.go is deleted — all
its contents now live in enterprise/coderd/chatd/.
