+++ enterprise/coderd/chatd/chatd.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"encoding/json"
 	"net/http"
+	"net/url"
 	"time"
 
 	"github.com/google/uuid"
@@ -15,22 +16,76 @@ import (
 	"github.com/coder/coder/v2/coderd/database/db2sdk"
 	coderdpubsub "github.com/coder/coder/v2/coderd/pubsub"
 	"github.com/coder/coder/v2/codersdk"
+	"github.com/coder/websocket"
 )
 
-// RemotePartsProvider returns a snapshot and live stream of
-// message_part events from the replica that is running the chat.
-// Called when the chat is actively running on a different replica.
-type RemotePartsProvider func(
+// RelaySourceHeader marks replica-relayed stream requests.
+const RelaySourceHeader = "X-Coder-Relay-Source-Replica"
+
+const (
+	authorizationHeader = "Authorization"
+	cookieHeader        = "Cookie"
+)
+
+// Config holds the dependencies for multi-replica chat
+// subscription. ReplicaIDFn is called lazily because the
+// replica ID may not be known at construction time.
+//
+// DialerFn, when set, overrides the default WebSocket relay
+// dialer. This is used in tests to inject mock relay behavior
+// without requiring real HTTP servers.
+type Config struct {
+	ResolveReplicaAddress func(context.Context, uuid.UUID) (string, bool)
+	ReplicaHTTPClient     *http.Client
+	ReplicaIDFn           func() uuid.UUID
+	DialerFn              func(
+		ctx context.Context,
+		chatID uuid.UUID,
+		workerID uuid.UUID,
+		requestHeader http.Header,
+	) (
+		snapshot []codersdk.ChatStreamEvent,
+		parts <-chan codersdk.ChatStreamEvent,
+		cancel func(),
+		err error,
+	)
+}
+
+// dial returns the dialer function to use for relay connections.
+// If DialerFn is set (e.g. in tests), it takes precedence.
+// Otherwise, dialRelay is used with the real Config dependencies.
+// Returns nil when no relay capability is configured.
+func (c Config) dial() func(
 	ctx context.Context,
 	chatID uuid.UUID,
 	workerID uuid.UUID,
 	requestHeader http.Header,
 ) (
-	snapshot []codersdk.ChatStreamEvent,
-	parts <-chan codersdk.ChatStreamEvent,
-	cancel func(),
-	err error,
-)
+	[]codersdk.ChatStreamEvent,
+	<-chan codersdk.ChatStreamEvent,
+	func(),
+	error,
+) {
+	if c.DialerFn != nil {
+		return c.DialerFn
+	}
+	if c.ResolveReplicaAddress == nil {
+		return nil
+	}
+	return func(
+		ctx context.Context,
+		chatID uuid.UUID,
+		workerID uuid.UUID,
+		requestHeader http.Header,
+	) (
+		[]codersdk.ChatStreamEvent,
+		<-chan codersdk.ChatStreamEvent,
+		func(),
+		error,
+	) {
+		return dialRelay(ctx, chatID, workerID, requestHeader, c)
+	}
+}
 
 // NewMultiReplicaSubscribeFn returns a SubscribeFn that merges local
 // message_part events, relay (remote-replica) message_part events,
@@ -38,7 +93,7 @@ type RemotePartsProvider func(
 // This captures all relay/pubsub merge logic that enterprise
 // multi-replica deployments require.
 func NewMultiReplicaSubscribeFn(
-	provider RemotePartsProvider,
+	cfg Config,
 ) osschatd.SubscribeFn {
 	return func(ctx context.Context, params osschatd.SubscribeMultiReplicaParams) (<-chan codersdk.ChatStreamEvent, func()) {
 		chatID := params.ChatID
@@ -59,8 +114,8 @@ func NewMultiReplicaSubscribeFn(
 		if params.Chat.Status == database.ChatStatusRunning &&
 			params.Chat.WorkerID.Valid &&
 			params.Chat.WorkerID.UUID != params.WorkerID &&
-			provider != nil {
-			snapshot, parts, cancel, err := provider(ctx, chatID, params.Chat.WorkerID.UUID, requestHeader)
+			cfg.dial() != nil {
+			snapshot, parts, cancel, err := cfg.dial()(ctx, chatID, params.Chat.WorkerID.UUID, requestHeader)
 			if err == nil {
 				relayCancel = cancel
 				relayParts = parts
@@ -136,7 +191,7 @@ func NewMultiReplicaSubscribeFn(
 		// goroutine and delivers the result on relayReadyCh so the
 		// main select loop is never blocked by network I/O.
 		openRelayAsync := func(workerID uuid.UUID) {
-			if provider == nil {
+			if cfg.dial() == nil {
 				return
 			}
 			closeRelay()
@@ -147,7 +202,7 @@ func NewMultiReplicaSubscribeFn(
 			dialCtx, dialCancel = context.WithCancel(ctx)
 			expectedWorkerID = workerID
 			go func() {
-				snapshot, parts, cancel, err := provider(dialCtx, chatID, workerID, requestHeader)
+				snapshot, parts, cancel, err := cfg.dial()(dialCtx, chatID, workerID, requestHeader)
 				if err != nil {
 					logger.Warn(ctx, "failed to open relay for message parts",
 						slog.F("chat_id", chatID),
@@ -202,7 +257,7 @@ func NewMultiReplicaSubscribeFn(
 		// loop can re-check chat status and reopen the relay
 		// without spinning in a tight loop.
 		scheduleRelayReconnect := func() {
-			if provider == nil {
+			if cfg.dial() == nil {
 				return
 			}
 			if reconnectTimer != nil {
@@ -444,18 +499,168 @@ func NewMultiReplicaSubscribeFn(
 			}()
 		}
 
-			// The cancel function only tears down the upstream
-			// sources (local stream, pubsub subscription). The
-			// merge goroutine owns all relay state (reconnectTimer,
-			// relayCancel, dialCancel, etc.) and cleans it up via
-			// its defer closeRelay() when ctx is cancelled.
-			cancel := func() {
-				for _, cancelFn := range allCancels {
-					if cancelFn != nil {
-						cancelFn()
-					}
+		// The cancel function only tears down the upstream
+		// sources (local stream, pubsub subscription). The
+		// merge goroutine owns all relay state (reconnectTimer,
+		// relayCancel, dialCancel, etc.) and cleans it up via
+		// its defer closeRelay() when ctx is cancelled.
+		cancel := func() {
+			for _, cancelFn := range allCancels {
+				if cancelFn != nil {
+					cancelFn()
 				}
 			}
+		}
 		return mergedEvents, cancel
 	}
+}
+
+// dialRelay opens a WebSocket relay connection to the replica
+// identified by workerID and returns a snapshot of buffered
+// message_part events plus a live channel of subsequent events.
+func dialRelay(
+	ctx context.Context,
+	chatID uuid.UUID,
+	workerID uuid.UUID,
+	requestHeader http.Header,
+	cfg Config,
+) (
+	snapshot []codersdk.ChatStreamEvent,
+	parts <-chan codersdk.ChatStreamEvent,
+	cancel func(),
+	err error,
+) {
+	address, ok := cfg.ResolveReplicaAddress(ctx, workerID)
+	if !ok {
+		return nil, nil, nil, xerrors.New("worker replica not found")
+	}
+
+	baseURL, err := url.Parse(address)
+	if err != nil {
+		return nil, nil, nil, xerrors.Errorf("parse relay address %q: %w", address, err)
+	}
+	replicaID := cfg.ReplicaIDFn()
+	relayCtx, relayCancel := context.WithCancel(ctx)
+	sdkClient := codersdk.New(baseURL)
+	sdkClient.HTTPClient = cfg.ReplicaHTTPClient
+	sdkClient.SessionTokenProvider = relayHeaderTokenProvider{
+		header: relayHeaders(requestHeader, replicaID),
+	}
+	sourceEvents, sourceStream, err := sdkClient.StreamChat(relayCtx, chatID)
+	if err != nil {
+		relayCancel()
+		return nil, nil, nil, xerrors.Errorf("dial relay stream: %w", err)
+	}
+
+	snapshot = make([]codersdk.ChatStreamEvent, 0, 100)
+
+	// Wait briefly for the first event to handle the common
+	// case where the remote side has buffered parts but hasn't
+	// flushed them to the WebSocket yet.
+	const drainTimeout = time.Second
+	drainTimer := time.NewTimer(drainTimeout)
+	defer drainTimer.Stop()
+
+drainInitial:
+	for len(snapshot) < cap(snapshot) {
+		select {
+		case <-relayCtx.Done():
+			_ = sourceStream.Close()
+			relayCancel()
+			return nil, nil, nil, xerrors.Errorf("dial relay stream: %w", relayCtx.Err())
+		case event, ok := <-sourceEvents:
+			if !ok {
+				break drainInitial
+			}
+			if event.Type != codersdk.ChatStreamEventTypeMessagePart {
+				continue
+			}
+			snapshot = append(snapshot, event)
+			// After getting the first event, switch to
+			// non-blocking drain for remaining buffered events.
+			drainTimer.Stop()
+			drainTimer.Reset(0)
+		case <-drainTimer.C:
+			break drainInitial
+		}
+	}
+
+	events := make(chan codersdk.ChatStreamEvent, 128)
+
+	go func() {
+		defer close(events)
+		defer relayCancel()
+		defer func() {
+			_ = sourceStream.Close()
+		}()
+
+		// No need to re-send snapshot events — they're
+		// returned to the caller directly.
+		for {
+			select {
+			case <-relayCtx.Done():
+				return
+			case event, ok := <-sourceEvents:
+				if !ok {
+					return
+				}
+				if event.Type != codersdk.ChatStreamEventTypeMessagePart {
+					continue
+				}
+				select {
+				case events <- event:
+				case <-relayCtx.Done():
+					return
+				}
+			}
+		}
+	}()
+
+	cancelFn := func() {
+		relayCancel()
+		_ = sourceStream.Close()
+	}
+	return snapshot, events, cancelFn, nil
+}
+
+type relayHeaderTokenProvider struct {
+	header http.Header
+}
+
+func (p relayHeaderTokenProvider) AsRequestOption() codersdk.RequestOption {
+	return func(req *http.Request) {
+		for key, values := range p.header {
+			for _, value := range values {
+				req.Header.Add(key, value)
+			}
+		}
+	}
+}
+
+func (p relayHeaderTokenProvider) SetDialOption(opts *websocket.DialOptions) {
+	if opts.HTTPHeader == nil {
+		opts.HTTPHeader = make(http.Header)
+	}
+	for key, values := range p.header {
+		for _, value := range values {
+			opts.HTTPHeader.Add(key, value)
+		}
+	}
+}
+
+func (p relayHeaderTokenProvider) GetSessionToken() string {
+	return p.header.Get(codersdk.SessionTokenHeader)
+}
+
+func relayHeaders(source http.Header, replicaID uuid.UUID) http.Header {
+	header := make(http.Header)
+	if source != nil {
+		for _, key := range []string{codersdk.SessionTokenHeader, authorizationHeader, cookieHeader} {
+			for _, value := range source.Values(key) {
+				header.Add(key, value)
+			}
+		}
+	}
+	header.Set(RelaySourceHeader, replicaID.String())
+	return header
 }
